app/requirements.txt
Pillow
tkinter
pyinstaller